requirements.txt
configparser
python-telegram-bot
openai
google-api-python-client
redis
pymysql
requests
lxml
selenium